requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
opencv-python
ultralytics
numpy
python-multipart
gunicorn
gdown